requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.16.0
geographiclib==2.0
geopandas==1.0.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pycryptodome==3.21.0
PyJWT==2.10.0
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
